Add filter for question categories
